feat: 修改应用文件
