+++ apps/apps.json
@@ -1,4 +1,4 @@
 {
-  "version": "0.0.3",
-  "package": ""
+  "version": "0.0.6",
+  "package": "https://1panel.oss-cn-hangzhou.aliyuncs.com/apps-0.0.6.zip"
 }
