--- apps/docker-compose.common.yml
@@ -1,4 +1,0 @@
-version: '3'
-networks:
-  1panel:
-    driver: bridge
