+++ apps/mysql5.7/metadata/logo.png
