+++ apps/mysql5.7/versions/5.7.39/conf/my.cnf
