+++ apps/mysql8.0/versions/8.0.30/config.json
