+++ apps/nginx/versions/1.23.1/conf/mime.types
