+++ apps/redis/metadata/logo.png
