+++ apps/redis/versions/7.0.5/docker-compose.yml
