+++ apps/wordpress/metadata/logo.png
