+++ apps/wordpress/versions/6.0.1/config.json
