+++ apps/wordpress/versions/6.0.2/config.json
