+++ backend/app/api/v1/app.go
@@ -4,9 +4,7 @@ import (
 	"github.com/1Panel-dev/1Panel/backend/app/api/v1/helper"
 	"github.com/1Panel-dev/1Panel/backend/app/dto"
 	"github.com/1Panel-dev/1Panel/backend/constant"
-	"github.com/1Panel-dev/1Panel/backend/global"
 	"github.com/gin-gonic/gin"
-	"reflect"
 )
 
 func (b *BaseApi) SearchApp(c *gin.Context) {
@@ -77,134 +75,4 @@ func (b *BaseApi) InstallApp(c *gin.Context) {
 	}
 
 	helper.SuccessWithData(c, install)
-}
-
-func (b *BaseApi) DeleteAppBackup(c *gin.Context) {
-	var req dto.AppBackupDeleteRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrBadRequest, constant.ErrTypeInvalidParams, err)
-		return
-	}
-	if err := appService.DeleteBackup(req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-		return
-	}
-
-	helper.SuccessWithData(c, nil)
-}
-
-func (b *BaseApi) SearchInstalled(c *gin.Context) {
-	var req dto.AppInstalledRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrBadRequest, constant.ErrTypeInvalidParams, err)
-		return
-	}
-	if !reflect.DeepEqual(req.PageInfo, dto.PageInfo{}) {
-		total, list, err := appService.PageInstalled(req)
-		if err != nil {
-			helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-			return
-		}
-
-		helper.SuccessWithData(c, dto.PageResult{
-			Items: list,
-			Total: total,
-		})
-	} else {
-		list, err := appService.SearchInstalled(req)
-		if err != nil {
-			helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-			return
-		}
-
-		helper.SuccessWithData(c, list)
-	}
-}
-
-func (b *BaseApi) SearchInstalledBackup(c *gin.Context) {
-	var req dto.AppBackupRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrBadRequest, constant.ErrTypeInvalidParams, err)
-		return
-	}
-	total, list, err := appService.PageInstallBackups(req)
-	if err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-		return
-	}
-
-	helper.SuccessWithData(c, dto.PageResult{
-		Items: list,
-		Total: total,
-	})
-}
-
-func (b *BaseApi) OperateInstalled(c *gin.Context) {
-	var req dto.AppInstallOperate
-	if err := c.ShouldBindJSON(&req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrBadRequest, constant.ErrTypeInvalidParams, err)
-		return
-	}
-	if err := appService.OperateInstall(req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-		return
-	}
-
-	helper.SuccessWithData(c, nil)
-}
-
-func (b *BaseApi) SyncInstalled(c *gin.Context) {
-	if err := appService.SyncAllInstalled(); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-		return
-	}
-	helper.SuccessWithData(c, "")
-}
-
-func (b *BaseApi) GetServices(c *gin.Context) {
-
-	key := c.Param("key")
-	services, err := appService.GetServices(key)
-	if err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-		return
-	}
-
-	helper.SuccessWithData(c, services)
-}
-
-func (b *BaseApi) GetUpdateVersions(c *gin.Context) {
-
-	appInstallId, err := helper.GetIntParamByKey(c, "appInstallId")
-	if err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrBadRequest, constant.ErrTypeInternalServer, nil)
-		return
-	}
-
-	versions, err := appService.GetUpdateVersions(appInstallId)
-	if err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-		return
-	}
-
-	helper.SuccessWithData(c, versions)
-}
-
-func (b *BaseApi) ChangeAppPort(c *gin.Context) {
-	var req dto.PortUpdate
-	if err := c.ShouldBindJSON(&req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrBadRequest, constant.ErrTypeInvalidParams, err)
-		return
-	}
-	if err := global.VALID.Struct(req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrBadRequest, constant.ErrTypeInvalidParams, err)
-		return
-	}
-
-	if err := appService.ChangeAppPort(req); err != nil {
-		helper.ErrorWithDetail(c, constant.CodeErrInternalServer, constant.ErrTypeInternalServer, err)
-		return
-	}
-
-	helper.SuccessWithData(c, nil)
 }
