+++ backend/app/api/v1/entry.go
@@ -11,7 +11,8 @@ var ApiGroupApp = new(ApiGroup)
 var (
 	authService = service.ServiceGroupApp.AuthService
 
-	appService = service.ServiceGroupApp.AppService
+	appService        = service.ServiceGroupApp.AppService
+	appInstallService = service.ServiceGroupApp.AppInstallService
 
 	containerService       = service.ServiceGroupApp.ContainerService
 	composeTemplateService = service.ServiceGroupApp.ComposeTemplateService
@@ -23,18 +24,18 @@ var (
 
 	cronjobService = service.ServiceGroupApp.CronjobService
 
-	hostService    = service.ServiceGroupApp.HostService
-	groupService   = service.ServiceGroupApp.GroupService
-	fileService    = service.ServiceGroupApp.FileService
+	hostService  = service.ServiceGroupApp.HostService
+	groupService = service.ServiceGroupApp.GroupService
+	fileService  = service.ServiceGroupApp.FileService
 
 	settingService = service.ServiceGroupApp.SettingService
 	backupService  = service.ServiceGroupApp.BackupService
 
 	operationService = service.ServiceGroupApp.OperationService
-	commandService         = service.ServiceGroupApp.CommandService
+	commandService   = service.ServiceGroupApp.CommandService
 
-	websiteGroupService    = service.ServiceGroupApp.WebsiteGroupService
-	websiteService         = service.ServiceGroupApp.WebsiteService
+	websiteGroupService       = service.ServiceGroupApp.WebsiteGroupService
+	websiteService            = service.ServiceGroupApp.WebsiteService
 	websiteDnsAccountService  = service.ServiceGroupApp.WebSiteDnsAccountService
 	websiteSSLService         = service.ServiceGroupApp.WebSiteSSLService
 	websiteAcmeAccountService = service.ServiceGroupApp.WebSiteAcmeAccountService
