+++ backend/app/repo/tag.go
@@ -3,26 +3,22 @@ package repo
 import (
 	"context"
 	"github.com/1Panel-dev/1Panel/backend/app/model"
-	"github.com/1Panel-dev/1Panel/backend/global"
-	"gorm.io/gorm"
 )
 
 type TagRepo struct {
 }
 
 func (t TagRepo) BatchCreate(ctx context.Context, tags []*model.Tag) error {
-	db := ctx.Value("db").(*gorm.DB)
-	return db.Create(&tags).Error
+	return getTx(ctx).Create(&tags).Error
 }
 
 func (t TagRepo) DeleteAll(ctx context.Context) error {
-	db := ctx.Value("db").(*gorm.DB)
-	return db.Where("1 = 1 ").Delete(&model.Tag{}).Error
+	return getTx(ctx).Where("1 = 1 ").Delete(&model.Tag{}).Error
 }
 
 func (t TagRepo) All() ([]model.Tag, error) {
 	var tags []model.Tag
-	if err := global.DB.Where("1 = 1 ").Find(&tags).Error; err != nil {
+	if err := getDb().Where("1 = 1 ").Find(&tags).Error; err != nil {
 		return nil, err
 	}
 	return tags, nil
@@ -30,7 +26,7 @@ func (t TagRepo) All() ([]model.Tag, error) {
 
 func (t TagRepo) GetByIds(ids []uint) ([]model.Tag, error) {
 	var tags []model.Tag
-	if err := global.DB.Where("id in (?)", ids).Find(&tags).Error; err != nil {
+	if err := getDb().Where("id in (?)", ids).Find(&tags).Error; err != nil {
 		return nil, err
 	}
 	return tags, nil
@@ -38,7 +34,7 @@ func (t TagRepo) GetByIds(ids []uint) ([]model.Tag, error) {
 
 func (t TagRepo) GetByKeys(keys []string) ([]model.Tag, error) {
 	var tags []model.Tag
-	if err := global.DB.Where("key in (?)", keys).Find(&tags).Error; err != nil {
+	if err := getDb().Where("key in (?)", keys).Find(&tags).Error; err != nil {
 		return nil, err
 	}
 	return tags, nil
@@ -46,7 +42,7 @@ func (t TagRepo) GetByKeys(keys []string) ([]model.Tag, error) {
 
 func (t TagRepo) GetByAppId(appId uint) ([]model.Tag, error) {
 	var tags []model.Tag
-	if err := global.DB.Where("id in (select tag_id from app_tags where app_id = ?)", appId).Find(&tags).Error; err != nil {
+	if err := getDb().Where("id in (select tag_id from app_tags where app_id = ?)", appId).Find(&tags).Error; err != nil {
 		return nil, err
 	}
 	return tags, nil
