+++ frontend/src/views/website/website/config/index.vue
@@ -2,9 +2,11 @@
     <LayoutContent :header="'网站设置'" :back-name="'Website'">
         <el-tabs v-model="index" @click="changeTab(index)">
             <el-tab-pane label="基本" name="basic">
-                <Basic :id="id" v-if="index === 'basic'"></Basic>
+                <Basic :key="id" :id="id" v-if="index === 'basic'"></Basic>
             </el-tab-pane>
-            <el-tab-pane label="安全">反代</el-tab-pane>
+            <el-tab-pane label="安全" name="safety">
+                <Safety :key="id" :id="id" v-if="index === 'safety'"></Safety>
+            </el-tab-pane>
             <el-tab-pane label="备份">反代</el-tab-pane>
             <el-tab-pane label="源文">反代</el-tab-pane>
         </el-tabs>
@@ -15,6 +17,7 @@
 import LayoutContent from '@/layout/layout-content.vue';
 import { onMounted, ref } from 'vue';
 import Basic from './basic/index.vue';
+import Safety from './safety/index.vue';
 import router from '@/routers';
 
 const props = defineProps({
@@ -36,9 +39,7 @@ const changeTab = (index: string) => {
 };
 
 onMounted(() => {
+    index.value = props.tab;
     id.value = Number(props.id);
-    if (props.tab !== index.value) {
-        index.value = props.tab;
-    }
 });
 </script>
