+++ frontend/src/views/website/website/config/safety/index.vue
@@ -1,0 +1,26 @@
+<template>
+    <el-tabs tab-position="left" type="border-card" v-model="index">
+        <el-tab-pane :label="'CC 防护'" :id="id"></el-tab-pane>
+        <el-tab-pane :label="'白名单'"></el-tab-pane>
+        <el-tab-pane :label="'黑名单'"></el-tab-pane>
+        <el-tab-pane :label="'并发限制'"></el-tab-pane>
+        <el-tab-pane :label="'请求限制'"></el-tab-pane>
+        <el-tab-pane :label="'禁止爬虫'"></el-tab-pane>
+    </el-tabs>
+</template>
+<script lang="ts" setup>
+import { computed, ref } from 'vue';
+
+const props = defineProps({
+    id: {
+        type: Number,
+        default: 0,
+    },
+});
+
+const id = computed(() => {
+    return props.id;
+});
+
+let index = ref('0');
+</script>
