+++ frontend/src/views/website/website/index.vue
@@ -4,8 +4,8 @@
             <template #toolbar>
                 <el-button type="primary" plain @click="openCreate">{{ $t('commons.button.create') }}</el-button>
                 <el-button type="primary" plain @click="openGroup">{{ $t('website.group') }}</el-button>
-                <el-button type="primary" plain>{{ '修改默认页' }}</el-button>
-                <el-button type="primary" plain>{{ '默认站点' }}</el-button>
+                <!-- <el-button type="primary" plain>{{ '修改默认页' }}</el-button>
+                <el-button type="primary" plain>{{ '默认站点' }}</el-button> -->
             </template>
             <el-table-column :label="$t('commons.table.name')" fix show-overflow-tooltip prop="primaryDomain">
                 <template #default="{ row }">
@@ -13,9 +13,9 @@
                 </template>
             </el-table-column>
             <el-table-column :label="$t('commons.table.status')" prop="status"></el-table-column>
-            <el-table-column :label="'备份'" prop="backup"></el-table-column>
+            <!-- <el-table-column :label="'备份'" prop="backup"></el-table-column> -->
             <el-table-column :label="'备注'" prop="remark"></el-table-column>
-            <el-table-column :label="'SSL证书'" prop="ssl"></el-table-column>
+            <!-- <el-table-column :label="'SSL证书'" prop="ssl"></el-table-column> -->
             <fu-table-operations
                 :ellipsis="1"
                 :buttons="buttons"
@@ -73,8 +73,10 @@ const openConfig = (id: number) => {
 
 const buttons = [
     {
-        label: '设置',
-        click: open,
+        label: i18n.global.t('website.config'),
+        click: function (row: WebSite.WebSite) {
+            openConfig(row.id);
+        },
     },
     {
         label: i18n.global.t('app.delete'),
